+++ Moda.Web/src/moda.web.reactclient/src/app/planning/planning-intervals/[key]/layout.tsx
@@ -1,5 +1,6 @@
 'use client'
 
+import { useGetPlanningIntervalByKey } from '@/src/services/queries/planning-queries'
 import { Layout, Menu, MenuProps } from 'antd'
 import Link from 'next/link'
 import { useMemo, useState } from 'react'
@@ -13,8 +14,16 @@ const PlanninIntervalLayout = ({
   params: any
   children: React.ReactNode
 }) => {
+  const { data: planningIntervalData } = useGetPlanningIntervalByKey(params.key)
+
   const items: MenuProps['items'] = useMemo(
     () => [
+      {
+        label: `Planning Interval: ${planningIntervalData?.name}`,
+        key: 'pi-name',
+        type: 'text',
+        disabled: true,
+      },
       {
         label: (
           <Link href={`/planning/planning-intervals/${params.key}`}>
@@ -31,10 +40,14 @@ const PlanninIntervalLayout = ({
         ),
         key: 'pi-plan-review',
       },
-      //   {
-      //     label: 'Objectives',
-      //     key: 'pi-objectives',
-      //   },
+      {
+        label: (
+          <Link href={`/planning/planning-intervals/${params.key}/objectives`}>
+            Objectives
+          </Link>
+        ),
+        key: 'pi-objectives',
+      },
       //   {
       //     label: 'Risks',
       //     key: 'pi-risks',
@@ -56,7 +69,7 @@ const PlanninIntervalLayout = ({
         ],
       },
     ],
-    [params.key],
+    [planningIntervalData?.name, params.key],
   )
 
   return (
