+++ Moda.Web/src/moda.web.reactclient/src/app/planning/planning-intervals/[key]/objectives/loading.tsx
@@ -3,10 +3,10 @@
 import PageTitle from '@/src/app/components/common/page-title'
 import { Skeleton } from 'antd'
 
-export default function PlanningIntervalDetailsLoading() {
+export default function PlanningIntervalObjectivesPageLoading() {
   return (
     <>
-      <PageTitle title="PI Details" />
+      <PageTitle title="PI Objectives" />
       <Skeleton active />
     </>
   )
